fix: move extensions to businesslogic
fix: extension location
