+++ Livefront.BusinessLogic/Extensions/ReferralLinkExtensions.cs
@@ -1,16 +1,24 @@
-using Livefront.Referrals.API.Models;
+using Livefront.BusinessLogic.Models;
 using Livefront.Referrals.DataAccess.Models;
 using Livefront.Referrals.DataAccess.Models.DeeplinkApi.Models;
 
-namespace Livefront.Referrals.API.Extensions;
+namespace Livefront.BusinessLogic.Extensions;
 
 public static class ReferralLinkExtensions
 {
-    public static ReferralLinkDTO? ToReferralLinkDto(this ReferralLink? referralLink)
+    public static ReferralLinkDTO ToReferralLinkDto(this ReferralLink referralLink)
     {
-        if(referralLink == null)
+        if (referralLink == null)
         {
-            return null;
+            throw new ArgumentNullException(nameof(referralLink), "Referral link cannot be null.");
+        }
+        if(string.IsNullOrWhiteSpace(referralLink.BaseDeepLink))
+        {
+            throw new ArgumentNullException(nameof(referralLink), "Referral link cannot be null or empty.");
+        }
+        if(referralLink.ExpirationDate == default)
+        {
+            throw new ArgumentNullException(nameof(referralLink), "Referral link expiration date must be valid.");
         }
         
         return new ReferralLinkDTO
