+++ Livefront.BusinessLogic/Models/ReferralLinkDTO.cs
@@ -1,7 +1,6 @@
 using System.Text.Json.Serialization;
-using Livefront.Referrals.DataAccess.Models;
 
-namespace Livefront.Referrals.API.Models;
+namespace Livefront.BusinessLogic.Models;
 
 public class ReferralLinkDTO
 {
