+++ Livefront.Referrals.UnitTests/BusinessLogic/Extensions/ReferralExtensionsTests/WhenCallingToReferralDto.cs
@@ -1,0 +1,6 @@
+namespace Livefront.Referrals.UnitTests.BusinessLogic.Extensions.ReferralExtensionsTests;
+
+public class WhenCallingToReferralDto
+{
+    
+}
